Fix fee state management
