+++ src/app/screens/swap/swapConfirmation/useConfirmSwap.tsx
@@ -2,6 +2,7 @@ import { Currency, SponsoredTxErrorCode, SponsoredTxError } from 'alex-sdk';
 import { useNavigate } from 'react-router-dom';
 import { SwapToken } from '@screens/swap/useSwap';
 import useWalletSelector from '@hooks/useWalletSelector';
+import { microstacksToStx } from '@secretkeylabs/xverse-core';
 import {
   broadcastSignedTransaction,
   signTransaction,
@@ -12,6 +13,9 @@ import useNetworkSelector from '@hooks/useNetwork';
 import { ApiResponseError } from '@secretkeylabs/xverse-core/types';
 import { TokenImageProps } from '@components/tokenImage';
 import { useAlexSponsoredTransaction } from '../useAlexSponsoredTransaction';
+import { useState } from 'react';
+import BigNumber from 'bignumber.js';
+import { useCurrencyConversion } from '../useCurrencyConversion';
 
 export type SwapConfirmationInput = {
   from: Currency;
@@ -31,6 +35,7 @@ export type SwapConfirmationInput = {
 
 export type SwapConfirmationOutput = Omit<SwapConfirmationInput, 'unsignedTx'> & {
   onConfirm: () => Promise<void>;
+  onFeeUpdate: (settingFee: bigint) => void;
   unsignedTx: StacksTransaction; // deserialized StacksTransaction
 };
 
@@ -40,15 +45,27 @@ export function useConfirmSwap(input: SwapConfirmationInput): SwapConfirmationOu
   const { isSponsored, sponsorTransaction } = useAlexSponsoredTransaction(
     input.userOverrideSponsorValue,
   );
+  const { currencyToToken } = useCurrencyConversion();
   const navigate = useNavigate();
-  const unsignedTx = deserializeTransaction(input.unsignedTx);
+  const [unsignedTx, setUnsignedTx] = useState<StacksTransaction>(
+    deserializeTransaction(input.unsignedTx),
+  );
+  const [feeAmount, setFeeAmount] = useState(input.lpFeeAmount);
+  const [feeFiatAmount, setFeeFiatAmount] = useState(input.lpFeeFiatAmount);
 
   return {
     ...input,
-    lpFeeAmount: isSponsored ? 0 : input.lpFeeAmount,
-    lpFeeFiatAmount: isSponsored ? 0 : input.lpFeeFiatAmount,
+    lpFeeAmount: isSponsored ? 0 : feeAmount,
+    lpFeeFiatAmount: isSponsored ? 0 : feeFiatAmount,
     unsignedTx,
     userOverrideSponsorValue: input.userOverrideSponsorValue,
+    onFeeUpdate: (settingFee: bigint) => {
+      const fee = microstacksToStx(new BigNumber(settingFee));
+      unsignedTx.setFee(settingFee);
+      setUnsignedTx(unsignedTx);
+      setFeeAmount(Number(fee));
+      setFeeFiatAmount(currencyToToken(input.from, Number(fee))?.fiatAmount);
+    },
     onConfirm: async () => {
       const signed = await signTransaction(
         unsignedTx,
