+++ src/app/screens/swap/useSwap.tsx
@@ -9,16 +9,14 @@ import {
 import { useTranslation } from 'react-i18next';
 import useWalletSelector from '@hooks/useWalletSelector';
 import { TokenImageProps } from '@components/tokenImage';
-import { LoaderSize } from '@utils/constants';
 import { AlexSDK, Currency } from 'alex-sdk';
-import { ftDecimals } from '@utils/helper';
 import BigNumber from 'bignumber.js';
-import { getFiatEquivalent } from '@secretkeylabs/xverse-core/transactions';
 import { useNavigate } from 'react-router-dom';
 import { SwapConfirmationInput } from '@screens/swap/swapConfirmation/useConfirmSwap';
 import { AnchorMode, makeUnsignedContractCall, PostConditionMode } from '@stacks/transactions';
 import useStxPendingTxData from '@hooks/queries/useStxPendingTxData';
 import { useAlexSponsoredTransaction } from './useAlexSponsoredTransaction';
+import { useCurrencyConversion } from './useCurrencyConversion';
 
 const isNotNull = <T extends any>(t: T | null | undefined): t is T => t != null;
 
@@ -119,36 +117,11 @@ export function useSwap(): UseSwap {
   const navigate = useNavigate();
   const alexSDK = useState(() => new AlexSDK())[0];
   const { t } = useTranslation('translation', { keyPrefix: 'SWAP_SCREEN' });
-  const {
-    coins: supportedCoins = [],
-    coinsList: visibleCoins = [],
-    stxAvailableBalance,
-    stxBtcRate,
-    btcFiatRate,
-    stxAddress,
-    stxPublicKey,
-  } = useWalletSelector();
+  const { stxAddress, stxPublicKey } = useWalletSelector();
+  const { acceptableCoinList, currencyToToken } = useCurrencyConversion();
   const [userOverrideSponsorValue, setUserOverrideSponsorValue] = useState(true);
   const { isSponsored, isServiceRunning } = useAlexSponsoredTransaction(userOverrideSponsorValue);
   const { data: stxPendingTxData } = useStxPendingTxData();
-
-  const acceptableCoinList = supportedCoins
-    .filter((sc) => alexSDK.getCurrencyFrom(sc.contract) != null)
-    // TODO tim: remove this once alexsdk fix issue here
-    // https://github.com/alexgo-io/alex-sdk/issues/2
-    .filter((sc) => sc.contract !== 'SP3K8BC0PPEVCV7NZ6QSRWPQ2JE9E5B6N3PA0KBR9.brc20-db20')
-    .map<FungibleToken>((sc) => {
-      const ft = (visibleCoins || []).find((vc) => vc.principal === sc.contract);
-      return {
-        ...ft,
-        ...sc,
-        principal: sc.contract,
-        assetName: '',
-        total_sent: ft?.total_sent ?? '0',
-        total_received: ft?.total_received ?? '0',
-        balance: ft?.balance ?? '0',
-      };
-    });
 
   const [inputAmount, setInputAmount] = useState('');
   const [slippage, setSlippage] = useState(0.04);
@@ -160,40 +133,6 @@ export function useSwap(): UseSwap {
   });
 
   const fromAmount = Number.isNaN(Number(inputAmount)) ? undefined : Number(inputAmount);
-
-  function currencyToToken(currency?: Currency, amount?: number): SwapToken | undefined {
-    if (currency == null) {
-      return undefined;
-    }
-    if (currency === Currency.STX) {
-      return {
-        balance: Number(microstacksToStx(BigNumber(stxAvailableBalance) as any)),
-        image: { token: 'STX', size: 28, loaderSize: LoaderSize.SMALL },
-        name: 'STX',
-        amount,
-        fiatAmount:
-          amount != null
-            ? Number(getFiatEquivalent(amount, 'STX', stxBtcRate as any, btcFiatRate as any))
-            : undefined,
-      };
-    }
-    const token = acceptableCoinList.find(
-      (c) => alexSDK.getCurrencyFrom(c.principal) === currency,
-    )!;
-    if (token == null) {
-      return undefined;
-    }
-    return {
-      amount,
-      image: { fungibleToken: token, size: 28, loaderSize: LoaderSize.SMALL },
-      name: (token.ticker ?? token.name).toUpperCase(),
-      balance: Number(ftDecimals(token.balance, token.decimals ?? 0)),
-      fiatAmount:
-        amount != null
-          ? Number(getFiatEquivalent(amount, 'FT', stxBtcRate as any, btcFiatRate as any, token))
-          : undefined,
-    };
-  }
 
   function getCurrencyName(currency: Currency) {
     if (currency === Currency.STX) {
